+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 <!-- The next version number should be the version already set in pubspec.yaml -->
 
+## [1.13.1] - 2026-07-28
+### Changed
+- Acknowledge new "types" paramter in loadConversations API from Social BB [#5912](https://github.com/rokwire/illinois-app/issues/5912).
+
+## [1.13.0] - 2026-07-27
+### Changed
+- Upgrade to Flutter 3.44.4 [#5869](https://github.com/rokwire/illinois-app/issues/5869).
+- Conversation APIs in Social service extended to support group messaged [#5581](https://github.com/rokwire/illinois-app/issues/5581).
+- Implemented deleteConversation(s) APIs in Social BB  [#5581](https://github.com/rokwire/illinois-app/issues/5581).
+### Added
+- Added SetUtils.set helper [#5581](https://github.com/rokwire/illinois-app/issues/5581).
+- Added ListUtils.remove helper [#5581](https://github.com/rokwire/illinois-app/issues/5581).
+- Added entity caching feature in Content service [#5889](https://github.com/rokwire/illinois-app/issues/5889).
+- Created broadcastIndividualMessage API in Socail service, added minor helpers [#5767](https://github.com/rokwire/illinois-app/issues/5767).
+### Fixed
+- Fixed broadcastIndividualMessage response processing [#5767](https://github.com/rokwire/illinois-app/issues/5767).
+
 ## [1.12.9] - 2026-02-27
 ### Added
 - Added FlexUI.localeEvalParam helper [#5637](https://github.com/rokwire/illinois-app/issues/5637).